refactor: reorganize worker adapters and split ORB components
