+++ src/scaler/entry_points/worker_adapter_orb.py
@@ -1,6 +1,6 @@
 from scaler.config.section.orb_worker_adapter import ORBWorkerAdapterConfig
 from scaler.utility.logging.utility import setup_logger
-from scaler.worker_adapter.orb.worker_adapter import ORBWorkerAdapter
+from scaler.worker_adapter.drivers.orb_common.worker_adapter import ORBWorkerAdapter
 
 
 def main():
