+++ src/scaler/worker_adapter/drivers/orb_common/exception.py
