+++ src/scaler/worker_adapter/drivers/orb_ec2/ami/build.sh
