+++ src/scaler/worker_adapter/drivers/orb_ec2/ami/opengris-scaler.pkr.hcl
