+++ tests/scheduler/test_scaling.py
@@ -46,7 +46,7 @@ from scaler.scheduler.controllers.policies.simple_policy.scaling.capability_scal
 from scaler.utility.identifiers import ClientID, ObjectID, TaskID, WorkerID
 from scaler.utility.logging.utility import setup_logger
 from scaler.utility.network_util import get_available_tcp_port
-from scaler.worker_adapter.native import NativeWorkerAdapter
+from scaler.worker_adapter.drivers.native import NativeWorkerAdapter
 from tests.utility.utility import logging_test_name
